# src/main/java/com/quickreads/user/api/model/Item.java
package com.quickreads.user.api.model;

import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
@JsonInclude(JsonInclude.Include.NON_NULL)
public class Item {
	
	private String itemName;
	private String type;
	private String email;
	private String itemContent;
}