# src/main/java/com/quickreads/user/api/model/QuickreadsUserDetails.java
package com.quickreads.user.api.model;

import java.util.List;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class QuickreadsUserDetails {
	
	private QuickReadsUser quickreadsUser;
	private List<Item> items;

}