# swagger.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8085'
basePath: /
tags:
  - name: authentication-controller
    description: Authentication Controller
  - name: item-controller
    description: Item Controller
  - name: quick-reads-user-controller
    description: Quick Reads User Controller
paths:
  /v1/add/item:
    post:
      tags:
        - item-controller
      summary: createWelcome
      operationId: createWelcomeUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: item
          description: item
          required: true
          schema:
            $ref: '#/definitions/Item'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ItemResponse'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /v1/authenticate:
    post:
      tags:
        - authentication-controller
      summary: authenticateUser
      operationId: authenticateUserUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: loginRequest
          description: loginRequest
          required: true
          schema:
            $ref: '#/definitions/LoginRequest'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/LoginResponse'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /v1/createUser:
    post:
      tags:
        - quick-reads-user-controller
      summary: createNewUser
      operationId: createNewUserUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: quickReadsUser
          description: quickReadsUser
          required: true
          schema:
            $ref: '#/definitions/QuickReadsUser'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/QuickReadsUserResponse'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/v1/deleteUser/{userId}':
    delete:
      tags:
        - quick-reads-user-controller
      summary: deleteUser
      operationId: deleteUserUsingDELETE
      produces:
        - application/json
      parameters:
        - name: userId
          in: query
          description: userId
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/GenericResponse'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  '/v1/getUser/{userId}':
    get:
      tags:
        - quick-reads-user-controller
      summary: getUserDetails
      operationId: getUserDetailsUsingGET
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: userId
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/QuickreadsUserDetails'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/v1/item/{id}':
    get:
      tags:
        - item-controller
      summary: getItem
      operationId: getItemUsingGET
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Item'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /v1/items:
    get:
      tags:
        - item-controller
      summary: getAllItems
      operationId: getAllItemsUsingGET
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Item'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /v1/updateUser:
    put:
      tags:
        - quick-reads-user-controller
      summary: deleteUser
      operationId: deleteUserUsingPUT
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: quickReadsUser
          description: quickReadsUser
          required: true
          schema:
            $ref: '#/definitions/QuickReadsUser'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/GenericResponse'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
definitions:
  GenericResponse:
    type: object
    properties:
      responseMsg:
        type: string
    title: GenericResponse
  Item:
    type: object
    properties:
      email:
        type: string
      itemContent:
        type: string
      itemName:
        type: string
      type:
        type: string
    title: Item
  ItemResponse:
    type: object
    properties:
      details:
        type: string
      status:
        type: string
    title: ItemResponse
  LoginRequest:
    type: object
    properties:
      password:
        type: string
      userName:
        type: string
    title: LoginRequest
  LoginResponse:
    type: object
    properties:
      jwt:
        type: string
      validUser:
        type: boolean
    title: LoginResponse
  QuickReadsUser:
    type: object
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
      phNumber:
        type: string
      userType:
        type: string
    title: QuickReadsUser
  QuickReadsUserResponse:
    type: object
    properties:
      name:
        type: string
      status:
        type: string
      userName:
        type: string
      userType:
        type: string
    title: QuickReadsUserResponse
  QuickreadsUserDetails:
    type: object
    properties:
      items:
        type: array
        items:
          $ref: '#/definitions/Item'
      quickreadsUser:
        $ref: '#/definitions/QuickReadsUser'
    title: QuickreadsUserDetails